refactor(search): address review on BusGridMaps cache
Follow-up to the entropy-driver perf work, addressing PR review:

- Fix module layering cycle: move BusGridMaps out of ops::aod_grid into
  a new primitives::bus_grid_maps. primitives::lane_index (which owns the
  per-bus-group cache) now imports it from a sibling primitives module
  instead of reaching up into ops, so primitives no longer depends on ops.

- Guard the all-zones merge (item 2): the cache key drops zone_id via
  lanes_for_all_zones. Investigation showed this is safe for the move
  types that actually reach the cache -- SiteBus and WordBus are the only
  kinds LaneIndex registers as lanes, and a source's encoded location
  carries its zone_id/word_id/site_id, so distinct lanes never share a
  source key. ZoneBus (where a backward lane's source lives in the
  destination zone) would break that, but is never registered today, and
  all bundled/example specs have zone_buses=0 (the Gemini specs are
  single-zone). Added a debug_assert (insert_unique) on the source-keyed
  maps so a future ZoneBus surfaces loudly instead of silently resolving
  to a last-insert-wins over HashMap order. pos_to_src is deliberately
  left unguarded: distinct sources can legitimately share a physical
  position (the full.json fixture stacks every word at identical grid
  coordinates) -- a pre-existing, tolerated geometry property unrelated to
  the zone-merge hazard.

- Item 3 (shrink the maps) evaluated and declined: lane_endpoints shows a
  backward lane's source is the forward destination, so lane.{zone,word,
  site} != src.{zone,word,site} in general (and lane.zone != src.zone for
  ZoneBus backward). Lane reconstruction from a source is therefore
  invalid for backward lanes, so src_to_lane cannot be dropped; removing
  src_to_dst would push an endpoints() lookup into the hot path. Rationale
  documented on the BusGridMaps doc comment.

Behavior-preserving: the generated maps are byte-identical (the assert
only observes), the entropy fingerprint bench is unchanged, and the full
search test suite passes with debug assertions active.
